backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "flask-cors>=6.0.1",
    "gtts>=2.5.4",
    "langchain-openai>=0.3.31",
    "mcp-use>=1.3.10",
    "mcp[cli]>=1.13.1",
    "pre-commit>=4.3.0",
    "python-dotenv>=1.1.1",
]
